:recycle: consolehook -> console
