+++ .changeset/plenty-ways-exist.md
@@ -1,0 +1,5 @@
+---
+"vsit": minor
+---
+
+replace console with consolehook
